Divide email_processors module in to sub-package
